+++ Editor/AssetReferenceExporter.cs
@@ -213,22 +213,6 @@ namespace Com.Pamcha.CodaSync {
             
             return assets.ToArray();
         }
-
-        public static List<T> FindAssetsByType<T>(string folder) where T : UnityEngine.Object {
-            List<T> assets = new List<T>();
-            string[] classPath = typeof(T).ToString().Split('.');
-            string type = classPath[classPath.Length-1];
-            string[] guids = AssetDatabase.FindAssets($"t:{type}", new[] { folder });
-
-            for (int i = 0; i < guids.Length; i++) {
-                string assetPath = AssetDatabase.GUIDToAssetPath(guids[i]);
-                T asset = AssetDatabase.LoadAssetAtPath<T>(assetPath);
-                if (asset != null) {
-                    assets.Add(asset);
-                }
-            }
-            return assets;
-        }
         #endregion
     }
 }